Split npm-specific inputs from shared inputs
