+++ build-scan-shared/src/setup/input.ts
@@ -18,22 +18,6 @@ export function getDevelocityMavenExtensionVersion(): string {
     return sharedInput.getInput('develocity-maven-extension-version')
 }
 
-export function getDevelocityNpmAgentUrlOverride(): string {
-    return sharedInput.getInput('develocity-npm-agent-url-override')
-}
-
-export function getDevelocityNpmAgentVersion(): string {
-    return sharedInput.getInput('develocity-npm-agent-version')
-}
-
-export function getDevelocityNpmAgentInstallLocation(): string {
-    return sharedInput.getInput('develocity-npm-agent-install-location')
-}
-
-export function getDevelocityPacoteVersion(): string {
-    return sharedInput.getInput('develocity-pacote-version')
-}
-
 export function getCcudExtensionVersion(): string {
     return sharedInput.getInput('develocity-ccud-extension-version')
 }
